Better translation extraction.

+++ scripts/extract-translations.sh
@@ -8,16 +8,30 @@ else
   exit 1
 fi
 
+##### COMMANDLINE PARAMETERS #####
+LANGUAGE="$1"
+if [ -z "$LANGUAGE" ] || [ "$LANGUAGE" == 'en' ] ; then
+  LANGUAGE='fr'
+fi
+
 ##### EXTRACT TRANSLATIONS #####
 cd $DRUPAL_ROOT
 for site in sites/* ; do
   if [ -e $site/settings.php ] ; then
     cd $site
+
     MESSAGE "Extracting translations in $site..."
     INDENT
 
-    MESSAGE "Switching to Development mode..."
-    drush tadaa-switch dev >> $LOG_FILE 2>&1
+    # Check if DruTACT language module is enabled
+    MESSAGE "Checking for DruTACT language drutact_${LANGUAGE}..."
+    drush pm-list --type=module --status=enabled --no-core --pipe | grep drutact_${LANGUAGE} >> $LOG_FILE 2>&1
+    if [ $? != 0 ] ; then
+      ERROR "DruTACT language module (drutact_${LANGUAGE}) is not enabled."
+    fi
+
+    MESSAGE "Enabling the potx module..."
+    drush en potx -y >> $LOG_FILE 2>&1
 
     MESSAGE "Moving contrib libraries / modules / themes out of DruTACT profile..."
     mv ../../profiles/drutact/libraries ../all/libraries
@@ -40,8 +54,16 @@ for site in sites/* ; do
     drush rr >> $LOG_FILE 2>&1
     drush cc drush >> $LOG_FILE 2>&1
 
-    MESSAGE "Moving translations into DruTACT profile..."
-    mv ../../general.pot ../../profiles/drutact/translations/drutact.fr.po
+    MESSAGE "Disabling the potx module..."
+    drush dis potx -y >> $LOG_FILE 2>&1
+    drush pm-uninstall potx -y >> $LOG_FILE 2>&1
+
+    if [ -e ../../general.pot ] ; then
+      MESSAGE "Moving translations into DruTACT profile..."
+      mv ../../general.pot ../../profiles/drutact/translations/drutact.fr.po
+    else
+      ERROR "There was an error while extracting translations."
+    fi
 
     OUTDENT
     cd ../..
